Move stack generation to the UT #35
- Refactoring - separated stack generation and model building functionality to independent files.
- stack generator is ready to be moved elsewhere.
- UT :+1:
