+++ Opc.Ua.ModelCompiler/ModelCompilerAPI.cs
@@ -40,16 +40,4 @@ namespace OOI.ModelCompiler
       }
     }
   }
-
-  public static class ModelCompiler
-  {
-    public static void BuildModel(string outputDir, IModelGeneratorGenerate generateParameters, IModelGeneratorValidate validateParameters)
-    {
-      if (string.IsNullOrEmpty(outputDir))
-        throw new ArgumentOutOfRangeException("OutputDir", "Parameters cannot be null or empty");
-      ModelGenerator2 Generator = new ModelGenerator2();
-      Generator.ValidateAndUpdateIds(validateParameters);
-      Generator.Generate(generateParameters, outputDir);
-    }
-  }
 }
